server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
pymongo==4.6.0
werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
pydantic[email]==2.5.0